remove old main code and refactor
Refs #35
